+++ src/backend/config/KsnipConfig.h
@@ -29,6 +29,8 @@
 #include <QKeySequence>
 #include <QSettings>
 
+#include <kImageAnnotator/include/NumberUpdateMode.h>
+
 #include "KsnipConfigOptions.h"
 #include "src/common/enum/CaptureModes.h"
 #include "src/common/enum/SaveQualityMode.h"
@@ -125,6 +127,9 @@ public:
 
 	virtual bool switchToSelectToolAfterDrawingItem() const;
 	virtual void setSwitchToSelectToolAfterDrawingItem(bool enabled);
+
+	virtual bool numberToolSeedChangeUpdatesAllItems() const;
+	virtual void setNumberToolSeedChangeUpdatesAllItems(bool enabled);
 
 	virtual bool textBold() const;
 	virtual void setTextBold(bool bold);