+++ src/backend/config/KsnipConfigOptions.h
@@ -31,6 +31,7 @@ public:
 	static QString autoSaveNewCapturesString();
 	static QString rememberToolSelectionString();
 	static QString switchToSelectToolAfterDrawingItemString();
+	static QString numberToolSeedChangeUpdatesAllItemsString();
 	static QString useTabsString();
 	static QString autoHideTabsString();
 	static QString captureOnStartupString();